4.21*: Blog List Expansion, step 9

+++ controllers/blogs.js
@@ -34,8 +34,18 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const blogToDelete = await Blog.findById(request.params.id);
+  console.log("blogToDelete.user.toString()", blogToDelete.user.toString());
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  const user = await User.findById(decodedToken.id);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token invalid" });
+  } else if (blogToDelete.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: "unauthorized" });
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
